Moved api function to controller
