+++ controllers/controller.mjs
@@ -15,5 +15,18 @@ export async function serveCurrentData(team) {
         const currentData = await fetchLocations(team);
         return currentData;
     }
+};
+
+export async function sendTeamData(req, res) {
+    // Convert binary string to json and get the team page being viewed.
+    const jsonData = JSON.stringify(req.body);
+    const teamObject = JSON.parse(jsonData);
+    const team = teamObject.team;
+
+    // Get the current location data for the specified team.
+    const currentData = await serveCurrentData(team);
+    
+    // Send current location data as json.
+    res.json(currentData);
 }
